requirements.txt
apache-airflow
kaggle
pyspark
dotenv
apache-airflow-providers-postgres>=5.0.0
boto3